+++ README.md
@@ -1,88 +1,9 @@
 # mongoose-app
 Create Express API using Mongoose and more ...
 
-## Lab
-#### `1` Create a conditional that returns all users based off of `state`
+## Installation
 
-#### `2` Create a `Vehicle` model with the following:
-   - [ ] make
-   - [ ] model
-   - [ ] type
-   - [ ] vin
-   - [ ] color
-   - [ ] fuel
-
-#### `3` Make a `createRandomVehicles` function and test function
-   - [ ] commit message
+MongoDB Driver
 ```
-commit -m feat: add Vehicle model and createRandomVehicles function
-```
-
-#### `4` Create a `/vehicles` route
-
-#### `5` Create a `/vehicles/:field/:value` route
-   - [ ] Make conditionals with each field
-   - [ ] Add commit message
-```
-commit -m feat: add /vehicles and /vehicles/:field/:value route
-```
-
-#### `6` Create a `Company` model with the following:
-   - [ ] name
-   - [ ] address ( location )
-   - [ ] city
-   - [ ] state
-   - [ ] zipCode
-         
-#### `7` Make a `createRandomCompany` function
-   - [ ] commit message
-```
-commit -m feat: add Company model and createRandomCompany function
-```
-#### `8` Create a `/companies` route
-
-#### `9` Create a `/companies/:field/:value` route
-   - [ ] Make conditionals with each field
-   - [ ] Add commit message
-```
-commit -m feat: add /companies and /companies/:field/:value route
-```
-
-#### `10` Create a `Airline` model with the following:
-   - [ ] aircraftType
-   - [ ] airline
-   - [ ] airplane
-   - [ ] airport
-   - [ ] flightNumber
-   - [ ] recordLocator
-   - [ ] seat
-
-#### `11` Make a `createRandomAirlines` function and test
-
-#### `12` Create a `/airlines` route
-
-#### `13` Create a `/airlines/:field/:value` route
-  - [ ] Make conditionals with each field
-  - [ ] Add commit message
-
-```
-commit -m feat: add Airlines model and createRandomAirlines function
-```
-
-Create a `Commerce` model with the following:
-   - [ ] department
-   - [ ] price
-   - [ ] product
-   - [ ] productAdjective
-   - [ ] productDescription
-   - [ ] productMaterial
-   - [ ] productName
-
-Make a `createRandomCommerce` function
-commit message
-Create a `/commerce` route
-Create a `/commerce/:field/:value` route
-conditionals with each field
-```
-commit -m feat: add Commerce model and routes
+npm install mongodb
 ```
