+++ app.js
@@ -15,7 +15,7 @@ app.get('/', (req, res) => {
     return res.json({ message: 'Welcome to my API' });
 });
 
-app.use('/posts', require('./controllers/posts'));
+app.use('/medications', require('./controllers/medications'));
 app.use('/users', require('./controllers/users'));
 
 const PORT = process.env.PORT || 8000;
