+++ models/index.js
@@ -3,9 +3,9 @@ const mongoose = require('mongoose');
 
 // import all models
 const User = require('./user');
-const Post = require('./post');
-const Product = require('./product');
-const Order = require('./order');
+const Medication = require('./medication');
+const Dose = require('./dose');
+const Prescription = require('./prescription');
 
 console.log('mongo uri =>', process.env.MONGO_URI);
 
@@ -30,7 +30,7 @@ db.on('error', (err) => {
 
 module.exports = {
     User,
-    Post,
-    Product,
-    Order
-}
+    Prescription,
+    Medication,
+    Dose
+};
