+++ package.json
@@ -22,6 +22,7 @@
     "jsonwebtoken": "^9.0.1",
     "method-override": "^3.0.0",
     "mocha": "^10.2.0",
+    "mongodb": "^5.7.0",
     "mongoose": "^7.2.4",
     "passport": "^0.6.0",
     "passport-jwt": "^4.0.1",
